[refactor dist] Move `src/index.js` to `src/component.js`
