+++ src/component.js
